+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+- Payer Auth Setting Check error
+
+### Changed
+- Payer Auth updates
+
 ## [1.9.0] - 2022-12-06
 
 ### Changed